// server/src/app/enums/enumLanguageAPI.ts
/**
 * Enum que los posibles lenguajes que se pueden obtener de las cartas
 * 
 * @EN indica que la carta se puede desplegar en el idioma inglés
 * @FR indica que la carta se puede desplegar en el idioma francés
 * @ES indica que la carta se puede desplegar en el idioma español
 * @IT indica que la carta se puede desplegar en el idioma italiano
 * @PT indica que la carta se puede desplegar en el idioma portugués
 * @PT_BR indica que la carta se puede desplegar en el idioma portugués brasileño
 * @PT_PT indica que la carta se puede desplegar en el idioma portugués portugueño
 * @DE indica que la carta se puede desplegar en el idioma alemán
 * @NL indica que la carta se puede desplegar en el idioma holandés
 * @PL indica que la carta se puede desplegar en el idioma polaco
 * @RU indica que la carta se puede desplegar en el idioma ruso
 * @JA indica que la carta se puede desplegar en el idioma japonés
 * @KO indica que la carta se puede desplegar en el idioma coreano
 * @ZH_TW indica que la carta se puede desplegar en el idioma chino tradicional
 * @ID indica que la carta se puede desplegar en el idioma indonesio
 * @TH indica que la carta se puede desplegar en el idioma tailandés
 * @ZH_CN indica que la carta se puede desplegar en el idioma chino simplificado
 */
export enum ILanguage {
  EN = 'en', 
  FR = 'fr', 
  ES = 'es', 
  IT = 'it', 
  PT = 'pt', 
  PT_BR = 'pt-br', 
  PT_PT = 'pt-pt', 
  DE = 'de', 
  NL = 'nl', 
  PL = 'pl', 
  RU = 'ru', 
  JA = 'ja', 
  KO = 'ko', 
  ZH_TW = 'zh-tw', 
  ID = 'id', 
  TH = 'th', 
  ZH_CN = 'zh-cn' 
}